chore(message-card): organize test images (#5127)
