+++ src/message-card/__tests__/deliveryLargeStrawberries@3x.png
