+++ src/message-card/__tests__/earnerLargeRiderDriver@3x.png
