+++ src/message-card/__tests__/message-card-trailing-image.scenario.tsx
@@ -8,13 +8,13 @@ import * as React from 'react';
 import { colors } from '../../tokens';
 import { IMAGE_LAYOUT, BUTTON_KIND } from '../constants';
 import { MessageCard } from '..';
-import deliveryHeroItalianSvg from './images/deliveryHeroItalian.svg';
-import deliveryHeroItalianPng from './images/deliveryHeroItalian@3x.png';
-import deliveryLargeStrawberriesSvg from './images/deliveryLargeStrawberries.svg';
-import deliveryLargeStrawberriesPng from './images/deliveryLargeStrawberries@3x.png';
-import earnerLargeRiderDriverSvg from './images/earnerLargeRiderDriver.svg';
-import earnerLargeRiderDriverPng from './images/earnerLargeRiderDriver@3x.png';
-import veniceJpg from './images/venice.jpg';
+import deliveryHeroItalianSvg from './deliveryHeroItalian.svg';
+import deliveryHeroItalianPng from './deliveryHeroItalian@3x.png';
+import deliveryLargeStrawberriesSvg from './deliveryLargeStrawberries.svg';
+import deliveryLargeStrawberriesPng from './deliveryLargeStrawberries@3x.png';
+import earnerLargeRiderDriverSvg from './earnerLargeRiderDriver.svg';
+import earnerLargeRiderDriverPng from './earnerLargeRiderDriver@3x.png';
+import veniceJpg from './venice.jpg';
 
 export function Scenario() {
   const containerStyle = {
